requirements.txt
asgiref==3.5.2
black==22.10.0
click==8.1.3
dj-database-url==1.0.0
Django==4.1.3
django-cors-headers==3.13.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
mypy-extensions==0.4.3
pathspec==0.10.2
platformdirs==2.5.4
pycodestyle==2.10.0
PyJWT==2.6.0
python-dotenv==0.21.0
pytz==2022.6
sqlparse==0.4.3
tomli==2.0.1
